+++ src/binary64/tanh/tanh.c
@@ -263,7 +263,7 @@ double cr_tanh(double x){
   double t = v.f - 0x1.8p25;
 #endif
   int64_t i1 = (jt.u>>27)&0x3f, i0 = (jt.u>>33)&0x3f, ie = (int64_t)(jt.u<<13)>>52;
-  const b64u64_u sp = {.u = (1023 + ie)<<52};
+  const b64u64_u sp = {.u = (uint64_t)(1023 + ie)<<52};
   static const double ch[] = {0x1p+1, 0x1p+1, 0x1.55555557e54ffp+0, 0x1.55555553a12f4p-1};
   double t0h = t0[i0][1], t1h = t1[i1][1], th = t0h*t1h, tl;
   if(aix<0x400d76c8b4395810ull){ // |x| ~< 3.683
